requirements.txt
streamlit>=1.5.0
pillow>=8.2.0
redis>=4.3.4
googletrans>=3.1.0a0
pandas
pathlib
matplotlib
numpy
clip-by-openai
